requirements.txt
#devops
black==22.3.0
click==8.1.3 
pytest==7.1.3
pytest-cov==4.0.0
pylint==2.15.3
boto3==1.24.87
#web
fastapi==0.85.0
uvicorn==0.18.3

flask
openai==0.28
flask_bootstrap
python-dotenv


